Move logic into trees
