+++ change/@itwin-tree-widget-react-fc7f841b-2902-4cc8-ba93-07ebf2566a4e.json
@@ -1,0 +1,7 @@
+{
+  "type": "minor",
+  "comment": "Added error messages when stateless trees are filtered and too many results are found.",
+  "packageName": "@itwin/tree-widget-react",
+  "email": "33428304+jasdom@users.noreply.github.com",
+  "dependentChangeType": "patch"
+}
