+++ packages/itwin/tree-widget/src/components/trees/stateless/categories-tree/CategoriesTreeComponent.tsx
@@ -5,12 +5,10 @@
 
 import "../Tree.scss";
 import classNames from "classnames";
-import { Fragment, useState } from "react";
+import { Fragment } from "react";
 import { useActiveIModelConnection, useActiveViewport } from "@itwin/appui-react";
 import { IModelApp } from "@itwin/core-frontend";
-import { Flex, Text } from "@itwin/itwinui-react";
 import { UnifiedSelectionProvider } from "@itwin/presentation-hierarchies-react";
-import { TreeWidget } from "../../../../TreeWidget";
 import { TreeHeader } from "../../../tree-header/TreeHeader";
 import { AutoSizer } from "../../../utils/AutoSizer";
 import { HideAllButton, InvertAllButton, ShowAllButton } from "../../category-tree/CategoriesTreeButtons";
@@ -24,7 +22,6 @@ import type { SelectionStorage } from "@itwin/presentation-hierarchies-react";
 import type { CategoriesTreeHeaderButtonProps } from "../../category-tree/CategoriesTreeButtons";
 
 type StatelessCategoriesTreeProps = ComponentPropsWithoutRef<typeof StatelessCategoriesTree>;
-type StatelessCategoriesTreeError = Parameters<Required<StatelessCategoriesTreeProps>["onError"]>[0];
 
 interface StatelessCategoriesTreeComponentProps
   extends Pick<
@@ -57,7 +54,6 @@ function CategoriesTreeComponentImpl({
   selectionStorage,
   ...treeProps
 }: StatelessCategoriesTreeComponentProps & { iModel: IModelConnection; viewport: ScreenViewport }) {
-  const [error, setError] = useState<StatelessCategoriesTreeError | undefined>();
   const categories = useCategories(IModelApp.viewManager, iModel, viewport);
   const { filter, applyFilter } = useFiltering();
   const density = treeProps.density;
@@ -68,37 +64,10 @@ function CategoriesTreeComponentImpl({
     }
   };
 
-  const onFilterChanged = (newFilter: string) => {
-    error && setError(undefined);
-    applyFilter(newFilter);
-  };
-
-  const renderContent = (width: number, height: number) => {
-    if (error) {
-      return (
-        <Flex alignItems="center" justifyContent="center" flexDirection="column" style={{ width, height }}>
-          <FilterError error={error} />
-        </Flex>
-      );
-    }
-    return (
-      <StatelessCategoriesTree
-        {...treeProps}
-        imodel={iModel}
-        categories={categories}
-        activeView={viewport}
-        width={width}
-        height={height}
-        filter={filter}
-        onError={setError}
-      />
-    );
-  };
-
   return (
     <div className={classNames("tw-tree-with-header", density === "enlarged" && "enlarge")}>
       <UnifiedSelectionProvider storage={selectionStorage}>
-        <TreeHeader onFilterClear={() => onFilterChanged("")} onFilterStart={onFilterChanged} onSelectedChanged={() => {}} density={density}>
+        <TreeHeader onFilterClear={() => applyFilter("")} onFilterStart={applyFilter} onSelectedChanged={() => {}} density={density}>
           {headerButtons
             ? headerButtons.map((btn, index) => <Fragment key={index}>{btn({ viewport, categories, onFeatureUsed: onCategoriesTreeFeatureUsed })}</Fragment>)
             : [
@@ -114,13 +83,21 @@ function CategoriesTreeComponentImpl({
               ]}
         </TreeHeader>
         <div className="tw-tree-content">
-          <AutoSizer>{({ width, height }) => renderContent(width, height)}</AutoSizer>
+          <AutoSizer>
+            {({ width, height }) => (
+              <StatelessCategoriesTree
+                {...treeProps}
+                imodel={iModel}
+                categories={categories}
+                activeView={viewport}
+                width={width}
+                height={height}
+                filter={filter}
+              />
+            )}
+          </AutoSizer>
         </div>
       </UnifiedSelectionProvider>
     </div>
   );
-}
-
-function FilterError({ error }: { error: StatelessCategoriesTreeError }) {
-  return <Text>{TreeWidget.translate(`stateless.${error}`)}</Text>;
 }
